+++ index.ts
@@ -1,4 +1,7 @@
-export type {
-	AuthProvider,
-	AuthProviderConfig
+export {
+	type AuthProvider,
+	type AuthProviderConfig,
+	zodSession
 } from './lib/types'
+
+export { ClientAuthProvider, useSession } from './lib/provider'